requirements.txt
# Thư viện lõi cho Deep Learning và Mạng Neural
tensorflow>=2.10.0
keras>=2.10.0

# Xử lý ma trận và bảng biểu dữ liệu
numpy>=1.21.0
pandas>=1.3.0

# Xử lý hình ảnh và hiển thị
opencv-python>=4.6.0
Pillow>=9.0.0
matplotlib>=3.5.0

# Hỗ trợ chia tập dữ liệu và tính toán độ chính xác
scikit-learn>=1.0.0